+++ frontend/src/pages/index.js
@@ -1,0 +1,1 @@
+/** No import/export to allow for lazy loading */